Conflicts:
	build-tools/automation/azure-pipelines.yaml

+++ build-tools/xaprepare/xaprepare/Steps/Step_InstallGNUBinutils.cs
@@ -58,6 +58,8 @@ namespace Xamarin.Android.Prepare
 
 		bool CopyToDestination (Context context, string label, string sourceDir, string destinationDir, string osName = HostName, string[]? executableExtensions = null)
 		{
+			bool isWindows = osName == "windows";
+
 			Log.StatusLine ();
 			Log.StatusLine ($"Installing for {label}:");
 
@@ -74,7 +76,7 @@ namespace Xamarin.Android.Prepare
 
 			void CopyLibraries ()
 			{
-				if (osName == "windows") {
+				if (isWindows) {
 					return;
 				}
 
@@ -100,6 +102,23 @@ namespace Xamarin.Android.Prepare
 
 					Log.StatusLine ($"  {context.Characters.Bullet} Installing ", toolName, tailColor: ConsoleColor.White);
 					Utilities.CopyFile (toolSourcePath, toolDestinationPath);
+					File.WriteAllText (versionMarkerPath, DateTime.UtcNow.ToString ());
+
+					if (!isWindows) {
+						continue;
+					}
+
+					toolSourcePath = Path.ChangeExtension (toolSourcePath, ".pdb");
+					if (!File.Exists (toolSourcePath)) {
+						continue;
+					}
+
+					toolDestinationPath = Path.ChangeExtension (toolDestinationPath, ".pdb");
+
+					Log.StatusLine ($"  {context.Characters.Bullet} Installing ", toolName, tailColor: ConsoleColor.White);
+					Utilities.CopyFile (toolSourcePath, toolDestinationPath);
+
+					versionMarkerPath = GetVersionMarker (toolDestinationPath);
 					File.WriteAllText (versionMarkerPath, DateTime.UtcNow.ToString ());
 				}
 			}
